requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
css-inline==0.13.0
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.0
packaging==23.2
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.31.0
sendgrid==6.11.0
starkbank-ecdsa==2.2.0
urllib3==2.2.1
Werkzeug==3.0.1